refactor - move around code
